implement new version of fsflags

+++ src/interrupts/syscalls/mod.rs
@@ -204,7 +204,7 @@ extern "C" fn syscall_2_open(args: &mut RegistersMini, _isf: &mut InterruptStack
     crate::debug!("syscall open mid");
     let fd = current_process
         .open_files
-        .create_file_table(path::Path::from(&path), args.rsi as usize)
+        .create_file_table(path::Path::from(&path), args.rsi)
         .into_u64();
     crate::debug!("syscall open end {}", fd);
     // Puts the fd into rax
